move carousel from home to its index
